Remove duplicated code - add abstract Validators
