+++ src/main/java/com/shape/shape_api/circle/validator/CircleV1Validator.java
@@ -1,28 +1,15 @@
 package com.shape.shape_api.circle.validator;
 
-import com.shape.shape_api.circle.model.Circle;
-import com.shape.shape_api.exception.InvalidEntityException;
-import com.shape.shape_api.shape.validator.ShapeValidator;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 
 @Component
-public class CircleV1Validator implements ShapeValidator<Circle> {
+public class CircleV1Validator extends AbstractCircleValidator {
 
     @Override
     public List<String> getRequiredParams() {
         return List.of("radius");
     }
 
-    @Override
-    public void validateEntity(Circle circle) {
-        if (circle == null) {
-            throw new InvalidEntityException("Circle must not be null");
-        }
-
-        if (circle.getRadius() == null) {
-            throw new InvalidEntityException("Radius must not be null");
-        }
-    }
 }
